Produtos Atualizados

+++ Java/ApoloInsightsBackEnd/src/main/java/school/sptech/ApoloInsightsBackEnd/repository/ProdutoRepository.java
@@ -3,6 +3,6 @@ package school.sptech.ApoloInsightsBackEnd.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import school.sptech.ApoloInsightsBackEnd.domain.Produto;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 
 }
